GEP-8197 Sync with Latest CXB UI Packages

* Merged in GEP-8059 (pull request #1800)

GEP-8059, Grab and drag multi timelines sync functionality isn't working

* GEP-8059, Grab and drag multi timelines sync functionality isn't working

* Merged development into GEP-8059
* Multi redirection resolved


Approved-by: Muhammad  Bilal
* Merged in B_GEP-7981_Default_Image_For_Thumbnail (pull request #1812)

Default image added for thumbnails.

* Default image added for thumbnails.

* Thumbnail default image changed.


Approved-by: Sufyan Jamil
* Merged in GEP-8066-GEP-8067 (pull request #1815)

GEP-8066-GEP-8067

* GEP-8066-GEP-8067

* Merged development into GEP-8066-GEP-8067

Approved-by: Sufyan Jamil
* Merged in GEP-6426 (pull request #1813)

GEP-6426

* GEP-8059, Grab and drag multi timelines sync functionality isn't working

* GEP-6426, Manage Units & Devices Template: few fields on primary device and unit settings tab is accepting negative values for time

* Merged development into GEP-6426

Approved-by: Sufyan Jamil
* Merged in GEP-633-unit-detail-image (pull request #1795)

GEP-6332-unit-detail-image-updated

* images of Bodyworn types updated in unit detail page

* debuggers and console removed

* Merged development into GEP-633-unit-detail-image
* Merged development into GEP-633-unit-detail-image
* remove double redirection


Approved-by: Sufyan Jamil
* Merged in GEP-8223_Images_Not_Showing_On_Case_TimeLine_Detail (pull request #1814)

Cases - Case Sharing - Adding Evidence (Images) to Shared Case Is Not Showing Case ID in Asset Lister

* Cases - Case Sharing - Adding Evidence (Images) to Shared Case Is Not Showing Case ID in Asset Lister


Approved-by: Ahsanullah Qureshi
* Merged in C_GEP_SaveReload (pull request #1817)

C GEP SaveReload

* icon

* one option

* Merge branch 'development' into C_GEP_SaveReload

* Assign user save reload stop

* Merge branch 'development' into C_GEP_SaveReload

* modal Close

* Check All

* Asset Log New

* Merge branch 'development' into C_GEP_SaveReload

* Merged development into C_GEP_SaveReload

Approved-by: Muhammad Waseem
* GEP-8197-Sync with Latest CXB UI Packages- Conflict resolved and error fixation.

* GEP-8197-Sync with Latest CXB UI Packages-Syncing with latest packages, styling and audio player fixation has been done.

* GEP-8197-Sync with Latest CXB UI Packages-error fixation.


Approved-by: Mehmood Usman

+++ packages/evm/src/Application/Admin/UploadPolicies/UploadPoliciesLister/UploadPoliciesTemplateActionMenu.tsx
@@ -15,12 +15,11 @@ import { SetupConfigurationAgent } from '../../../../utils/Api/ApiAgent';
 type Props = {
   selectedItems?: any;
   row?: any;
-  getRowData: () => void;
-  getSelectedData: () => void;
+  updateSelectedItems: () => void;
   onMessageShow: (isSuccess: boolean, message: string) => void;
 };
 
-const UploadPoliciesTemplateActionMenu: React.FC<Props> = ({ selectedItems, row, getRowData, getSelectedData, onMessageShow }) => {
+const UploadPoliciesTemplateActionMenu: React.FC<Props> = ({ selectedItems, row, updateSelectedItems, onMessageShow }) => {
   const { t } = useTranslation<string>();
   const history = useHistory();
   const [nondefault, setnondefault] = useState(false);
@@ -28,21 +27,29 @@ const UploadPoliciesTemplateActionMenu: React.FC<Props> = ({ selectedItems, row,
 
   const deleteUploadPolicies = () => {
     let eventIds: number[] = [];
-    if (Array.isArray(selectedItems) && selectedItems.length > 0) {
+    // if (Array.isArray(selectedItems) && selectedItems.length > 0) {
+    //   eventIds = selectedItems.map((data: any) => {
+    //     return data.id;
+    //   });
+    // }
+    // else
+    // {
+    //   eventIds.push(row.id);
+    // }
+    if(row && !selectedItems.includes(row)) {
+      eventIds.push(row.id);
+    }
+    else if(Array.isArray(selectedItems) && selectedItems.length > 0) {
       eventIds = selectedItems.map((data: any) => {
         return data.id;
       });
     }
-    else
-    {
-      eventIds.push(row.id);
-    }
+    if (eventIds.length > 0) {
       SetupConfigurationAgent.deleteAllUploadPoliciesTemplate(eventIds)
         .then(function (response: any) {
           let { AssignIdName, UnAssignsIds } = response;
           deleteUploadPoliciesHandler(AssignIdName, UnAssignsIds);
-          getRowData();
-          getSelectedData();
+          updateSelectedItems();
         })
         .catch(function (error) {
           if (error) {
@@ -50,7 +57,7 @@ const UploadPoliciesTemplateActionMenu: React.FC<Props> = ({ selectedItems, row,
             return error;
           }
         });
-      
+    }
 
     const deleteUploadPoliciesHandler = (AssignIdName: any, UnAssignsIds: any) => {
       if (AssignIdName.length > 0) {
@@ -112,7 +119,7 @@ const UploadPoliciesTemplateActionMenu: React.FC<Props> = ({ selectedItems, row,
         }
 
       >
-        {selectedItems.length <= 1 ? (
+        {(row && !selectedItems.includes(row)) || (selectedItems.length <= 1)  ? (
           <MenuItem onClick={openCreateUploadPolicyForm}>
             <Restricted moduleId={63}>
               <div className="crx-meu-content crx-spac "  >
